rename: correct wrong folder name
